test(ci): split the heredoc scanner tests into their own module
Adding the quoted-heredoc class pushed the scanner test module to 503
lines, three over the taste-lint ceiling, and the error-count ratchet
from issue #3779 blocked the push. Splitting is the honest fix: each
class already covers a distinct scanner feature, so the boundary was
there to be taken. Suppressing the rule for three lines was the wrong
trade.

The shared helpers stay in the original module and the new file imports
them, matching tests/ci/test_validation_scripts_are_reachable.py.

402 and 118 lines. 48 tests, unchanged.
